+++ src/data/types/nucleotides/translation.rs
@@ -12,6 +12,8 @@ pub trait Translate: NucleotidesReadable + Sealed {
         AminoAcids(translate_sequence(self.nucleotide_bytes()))
     }
 
+    /// Translates the DNA sequence to [`AminoAcids`] until the first stop codon
+    /// in the current reading frame is found.
     #[inline]
     #[must_use]
     fn translate_to_stop(&self) -> AminoAcids {
@@ -24,6 +26,25 @@ pub trait Translate: NucleotidesReadable + Sealed {
             }
         }
         AminoAcids(out)
+    }
+
+    /// Slides base by base over the sequence until the next translated `aa` is
+    /// found and returns that index (or `None` otherwise).
+    #[inline]
+    #[must_use]
+    fn find_next_aa(&self, aa: u8) -> Option<usize> {
+        let needle = aa.to_ascii_uppercase();
+        self.to_overlapping_aa_iter().position(|aa| aa == needle)
+    }
+
+    /// Slides codon by codon over the sequence (reading frame starting from 0)
+    /// until the next translated `aa` is found and returns that index (or
+    /// `None` otherwise).
+    #[inline]
+    #[must_use]
+    fn find_next_aa_in_frame(&self, aa: u8) -> Option<usize> {
+        let needle = aa.to_ascii_uppercase();
+        self.to_aa_iter().position(|aa| aa == needle).map(|x| x * 3)
     }
 
     /// Creates an iterator for [`AminoAcids`] translation.
@@ -40,13 +61,27 @@ pub trait Translate: NucleotidesReadable + Sealed {
     fn to_aa_iter_with(&self, partial_codon_encoding: u8) -> TranslatedNucleotidesIter {
         TranslatedNucleotidesIter::new_with(self.nucleotide_bytes(), partial_codon_encoding)
     }
+
+    /// Creates an iterator that translates amino acids base by base over all
+    /// reading frames.
+    #[inline]
+    #[must_use]
+    fn to_overlapping_aa_iter(&self) -> OverlappingCodonsIter {
+        OverlappingCodonsIter::new(self.nucleotide_bytes())
+    }
 }
 
 impl<T: NucleotidesReadable + Sealed> Translate for T {}
 
-/// Iterator for translating [`Nucleotides`] into [`AminoAcids`].
+/// Iterator for translating [`Nucleotides`] into [`AminoAcids`] codon by codon
+/// with a reading frame starting from 0.
+///
+/// This is created by [`to_aa_iter`] and [`to_aa_iter_with`]. To translate all
+/// codons regardless of the reading frame, see [`OverlappingCodonsIter`].
 ///
 /// [`Nucleotides`]: crate::prelude::Nucleotides
+/// [`to_aa_iter`]: Translate::to_aa_iter
+/// [`to_aa_iter_with`]: Translate::to_aa_iter_with
 pub struct TranslatedNucleotidesIter<'a> {
     codons:                 std::slice::Iter<'a, [u8; 3]>,
     has_remainder:          bool,
@@ -54,12 +89,18 @@ pub struct TranslatedNucleotidesIter<'a> {
 }
 
 impl<'a> TranslatedNucleotidesIter<'a> {
+    /// Create a new [`TranslatedNucleotidesIter`] using `~` to encode partial
+    /// codons.
     #[inline]
+    #[must_use]
     fn new(seq: &'a [u8]) -> Self {
         Self::new_with(seq, b'~')
     }
 
+    /// Create a new [`TranslatedNucleotidesIter`], using
+    /// `partial_codon_encoding` to encode partial codons.
     #[inline]
+    #[must_use]
     fn new_with(seq: &'a [u8], partial_codon_encoding: u8) -> Self {
         let (codons, has_remainder) = {
             let (chunks, remainder) = seq.as_chunks::<3>();
@@ -103,7 +144,46 @@ impl Iterator for TranslatedNucleotidesIter<'_> {
 impl ExactSizeIterator for TranslatedNucleotidesIter<'_> {}
 impl std::iter::FusedIterator for TranslatedNucleotidesIter<'_> {}
 
-/// A codon is considered *partial* if it has fewer than 3 IUPAC bases (non-gap).
+/// Iterator for translating [`Nucleotides`] into [`AminoAcids`] base by base
+/// over all reading frames.
+///
+/// This is created by [`to_overlapping_aa_iter`]. To translate non-overlapping
+/// codons in the current reading frame, see [`TranslatedNucleotidesIter`].
+///
+/// [`Nucleotides`]: crate::prelude::Nucleotides
+/// [`to_overlapping_aa_iter`]: Translate::to_overlapping_aa_iter
+pub struct OverlappingCodonsIter<'a> {
+    codons: std::slice::Windows<'a, u8>,
+}
+
+impl<'a> OverlappingCodonsIter<'a> {
+    /// Create a new [`OverlappingCodonsIter`] from the provided sequence.
+    #[inline]
+    #[must_use]
+    fn new(seq: &'a [u8]) -> Self {
+        Self { codons: seq.windows(3) }
+    }
+}
+
+impl Iterator for OverlappingCodonsIter<'_> {
+    type Item = u8;
+
+    #[inline]
+    fn next(&mut self) -> Option<Self::Item> {
+        self.codons.next().map(StdGeneticCode::translate_codon)
+    }
+
+    #[inline]
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.codons.size_hint()
+    }
+}
+
+impl ExactSizeIterator for OverlappingCodonsIter<'_> {}
+impl std::iter::FusedIterator for OverlappingCodonsIter<'_> {}
+
+/// A codon is considered *partial* if it has fewer than 3 IUPAC bases
+/// (non-gap).
 #[inline]
 #[must_use]
 fn is_partial_codon(codon: [u8; 3]) -> bool {
@@ -119,6 +199,7 @@ pub fn translate_sequence(s: &[u8]) -> Vec<u8> {
     TranslatedNucleotidesIter::new(s).collect::<Vec<_>>()
 }
 
+/// Provides functionality to retrieve codons from a nucleotides sequences.
 pub trait GetCodons: NucleotidesReadable + Sealed {
     /// Gets the bases grouped into codons as a slice of arrays, starting with
     /// the first base. Any trailing bases are included in the second tuple
@@ -188,6 +269,8 @@ impl GetCodons for Nucleotides {}
 impl GetCodons for NucleotidesView<'_> {}
 impl GetCodons for NucleotidesViewMut<'_> {}
 
+/// Provides functionality to obtain mutable references to codons within
+/// nucleotides sequences.
 pub trait GetCodonsMut: NucleotidesMutable + Sealed {
     /// Gets the bases grouped into codons as a mutable slice of arrays,
     /// starting with the first base. Any trailing bases are included in the